Update to ver0.9!
1. Relaxed the assumption of infinite parking space to generate more
realistic traffic. Now if there is no parking space in the current zone,
vehicles will cruising in the surrounding (~5km) area (see the
goCruising() function in the ElectricVehicle.class).

2. Formalized the usage of random seeds. Now the same random seed will
generate the exact same result in single thread mode. In multi-thread
mode, this promise fails as different zones can compete with each other
for empty vehicles and the order of updating different zones depends on
the computational speed.

3. In 1.0 we will add the traffic light simulation.